app/requirements.txt
bcrypt==3.2.2
sqlalchemy
psycopg2
sqlalchemy==2.0.31
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.1.10
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
sqlmodel==0.0.19
starlette==0.37.2
fastapi==0.111.0
uvicorn==0.30.1
pika==1.3.1
Jinja2
python-jose==3.4.0
passlib
pytest
pytest-asyncio
httpx
api-analytics
